renamed button for starting game
