--- app/components/cards/start-card.tsx
@@ -1,12 +1,0 @@
-import Button from "../button/button";
-
-export default function StartCard() {
-  return (
-    <div className="start-card">
-      <p className="text-lg text-center font-semibold">
-        Test your general knowledge with 10 random questions!
-      </p>
-      <Button />
-    </div>
-  );
-}
